requirements.txt
flask 
requests
pillow